\name{X}
\alias{X}
\docType{data}
\title{
Input data X
}
\description{
Simulated input data
}
\usage{data(X)}
\format{
  A data frame with 20 observations on the following 5 variables:
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{ss1_x2}}{a numeric vector}
    \item{\code{ss2_x2}}{a numeric vector}
    \item{\code{ss1_x3}}{a numeric vector}
    \item{\code{ss2_x3}}{a numeric vector}
   }
}
\details{
The data consist of simulated variables, including an individual-level covariate (x1) and two area-level covariates (x2, x3) available at two
different spatial scales (ss1, ss2).
}
\examples{
data(X)
}
\keyword{datasets}
